+++ tomcat-base/src/shared/main/java/eu/xenit/alfresco/tomcat/embedded/alfresco/tomcat/AlfrescoTomcatFactoryHelper.java
@@ -14,6 +14,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Properties;
 import java.util.logging.Logger;
+import org.apache.tomcat.util.net.SSLHostConfigCertificate;
 
 public class AlfrescoTomcatFactoryHelper {
 
@@ -66,9 +67,16 @@ public class AlfrescoTomcatFactoryHelper {
         );
 
         SSLHostConfig sslHostConfig = new SSLHostConfig();
-        sslHostConfig.setCertificateKeystoreFile(alfrescoConfiguration.getTomcatSSLKeystore());
-        sslHostConfig.setCertificateKeystorePassword(alfrescoConfiguration.getTomcatSSLKeystorePassword());
-        sslHostConfig.setCertificateKeystoreType("JCEKS");
+        // This is deprecated in Tomcat 9 and deleted from Tomcat V10.1!
+//            sslHostConfig.setCertificateKeystoreFile(alfrescoConfiguration.getTomcatSSLKeystore());
+//            sslHostConfig.setCertificateKeystorePassword(alfrescoConfiguration.getTomcatSSLKeystorePassword());
+//            sslHostConfig.setCertificateKeystoreType("JCEKS");
+        SSLHostConfigCertificate sslHostConfigCertificate = new SSLHostConfigCertificate(sslHostConfig, SSLHostConfigCertificate.Type.UNDEFINED);
+        sslHostConfigCertificate.setCertificateKeystoreFile(alfrescoConfiguration.getTomcatSSLKeystore());
+        sslHostConfigCertificate.setCertificateKeystorePassword(alfrescoConfiguration.getTomcatSSLKeystorePassword());
+        sslHostConfigCertificate.setCertificateKeystoreType("JCEKS");
+
+        sslHostConfig.addCertificate(sslHostConfigCertificate);
         sslHostConfig.setTruststoreFile(alfrescoConfiguration.getTomcatSSLTruststore());
         sslHostConfig.setTruststorePassword(alfrescoConfiguration.getTomcatSSLTruststorePassword());
         sslHostConfig.setTruststoreType("JCEKS");